requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
tzdata==2025.2
pandas==2.3.1
pytz==2025.2